Only create "related items" for new packages
Realted items often contain links and markup which is not available in
the meta.xml and is therefore changed after the initial import on the
CKAN system. To not lose these changes, this information should not be
touched again by the harvester after the initial import.

Maybe in a later version the markup information can be included in the
meta.xml and this decision might be revised, but for now we simply do
not change the related items if the package already exists.