Always use dataset name for metadata and diffs.
We create the dataset metadata files (in the diffs folder) in the gather stage, using the dataset name. We were then checking the metadata and creating diff files in the import stage, based on the package id (which comes from the harvest object guid). Thus we were never checking the old metadata against the new.