feat: Remove 'url' of upload resource
Previously, when creating a resource with url_type 'upload', we added the filename of the resource file as the resource url.
This was presumably so that when the file was uploaded by ckan/ckan/lib/uploader.py:ResourceUpload, the uploader could use the file extension to guess the mimetype of the file.
We used a custom validator in ckanext-stadtzh-theme that would not validate resource urls if resource url_type was 'upload'.
However, the ResourceUpload class has now been changed to validate the url before using it to guess the mimetype, so this no longer works.
If we don't enter a url here, the uploader will instead use the file extension of the uploaded file to guess the mimetype.
It also overwrites the resource url with the url of the uploaded file, so there is really no reason to keep setting the url here.