tests: Run on ubuntu-latest with Python 3
For the linting job, actions/setup-python@v3 has stopped supporting Python 2, so let's see if we can run flake8 on Python 3 and get it to pass.

For the testing job, there is no point sticking to ubuntu-20.04 so that we can use Python 2, as we are running all tests in a container with the right Python version anyway.